wip - updating switcherlist item
