+++ coral-component-shell/src/scripts/ShellSwitcherListItem.js
@@ -24,23 +24,7 @@ const CLASSNAME = '_coral-Shell-switcherList';
  @extends {HTMLElement}
  @extends {SelectList}
  */
-class ShellSwitcherList extends SelectList {
-
-    constructor(){
-      super();
-      this._delegateEvents({
-            'click coral-selectlist-item': '_onItemClick'
-      });
-
-    }
-
-    _onItemClick(event) {
-       const item = event.matchedTarget;
-       if (item.hasAttribute("href")) {
-       const href = item.getAttribute("href");
-       window.open(href, '_self');
-       }
-     }
+class ShellSwitcherList extends SelectList.Item {
 
     /** @ignore */
     render() {
