+++ coral-component-shell/src/templates/selectListSwitcher.html
@@ -1,1 +1,2 @@
-<div handle="container" class="_coral-Shell-selectList-container"></div>
+<coral-selectlist handle="container" class="_coral-Shell-selectList-container"></coral-selectlist>
+
